requirement.txt
streamlit
moviepy
requests
google-generativeai
langchain
langchain_google_genai

